chore(chat): decompose ChatCoordinator into focused delegates
Split the 735-line monolithic ChatCoordinator into three delegates
using Kotlin interface delegation, following the SessionController
pattern:

- FeedSyncDelegate: feed sync, DB observation, ChatSummary projection
- EventStreamDelegate: event stream, applyUpdate, gap tracking,
  reactions, typing indicators
- MessagingDelegate: per-chat messaging, read pointers, paging,
  notifications

ChatCoordinator becomes an interface with sub-interfaces
(FeedOperations, EventStreamOperations, MessagingOperations).
RealChatCoordinator is the thin shell that wires delegates via
Channel<Event> flows, handles lifecycle, and observes feature flags.
